Update OptionsPatternsGuidanceParameterless.csproj
This extension method you're looking for is not in the `Microsoft.Extensions.DependencyInjection` package, but rather that namespace. [See this article](https://docs.microsoft.com/en-us/dotnet/api/microsoft.extensions.dependencyinjection.optionsservicecollectionextensions.addoptions?view=dotnet-plat-ext-5.0). Notice the namespace and the assembly, you need to reference the `Microsoft.Extensions.Options` package, as it contributes to the `Microsoft.Extensions.DependencyInjection` namespace.